Filter control panel schema in settings (#3914)
* Filter control panel schema in settings

* Add doc entry
