+++ src/components/manage/Controlpanels/Controlpanel.jsx
@@ -17,6 +17,8 @@ import { toast } from 'react-toastify';
 import { Form, Icon, Toolbar, Toast } from '@plone/volto/components';
 import { updateControlpanel, getControlpanel } from '@plone/volto/actions';
 
+import config from '@plone/volto/registry';
+
 import saveSVG from '@plone/volto/icons/save.svg';
 import clearSVG from '@plone/volto/icons/clear.svg';
 
@@ -42,64 +44,6 @@ const messages = defineMessages({
     defaultMessage: 'Info',
   },
 });
-
-// Filters props.controlpanel.schema to only valid/relevant fields
-const filterSchema = (controlpanel) => {
-  const panelType = controlpanel['@id'].split('/').pop();
-
-  const unwantedSettings = {
-    language: ['display_flags', 'always_show_selector'],
-    search: ['enable_livesearch'],
-    site: [
-      'display_publication_date_in_byline',
-      'icon_visibility',
-      'thumb_visibility',
-      'no_thumbs_portlet',
-      'no_thumbs_lists',
-      'no_thumbs_summary',
-      'no_thumbs_tables',
-      'thumb_scale_portlet',
-      'thumb_scale_listing',
-      'thumb_scale_table',
-      'thumb_scale_summary',
-      'toolbar_position',
-      'toolbar_logo',
-      'default_page',
-    ],
-    editing: ['available_editors', 'default_editor', 'ext_editor'],
-    imaging: [
-      'highpixeldensity_scales',
-      'quality_2x',
-      'quality_3x',
-      'picture_variants',
-      'image_captioning',
-    ],
-  };
-
-  // Creates modified version of properties object
-  const newPropertiesObj = Object.fromEntries(
-    Object.entries(controlpanel.schema.properties).filter(
-      ([key, val]) => !(unwantedSettings[panelType] || []).includes(key),
-    ),
-  );
-  // Filters props.controlpanel.schema.fieldsets.fields to only valid/relevant fields
-  const filterFields = (fields) => {
-    return fields.filter(
-      (field) => !(unwantedSettings[panelType] || []).includes(field),
-    );
-  };
-  // Creates modified version of fieldsets array
-  const newFieldsets = controlpanel.schema.fieldsets.map((fieldset) => {
-    return { ...fieldset, fields: filterFields(fieldset.fields) };
-  });
-
-  // Returns clone of props.controlpanel.schema, with updated properties/fieldsets
-  return {
-    ...controlpanel.schema,
-    properties: newPropertiesObj,
-    fieldsets: newFieldsets,
-  };
-};
 
 /**
  * Controlpanel class.
@@ -205,6 +149,8 @@ class Controlpanel extends Component {
    * @returns {string} Markup for the component.
    */
   render() {
+    const { filterControlPanelsSchema } = config.settings;
+
     if (this.props.controlpanel) {
       return (
         <div id="page-controlpanel">
@@ -213,7 +159,7 @@ class Controlpanel extends Component {
             <Form
               ref={this.form}
               title={this.props.controlpanel.title}
-              schema={filterSchema(this.props.controlpanel)}
+              schema={filterControlPanelsSchema(this.props.controlpanel)}
               formData={this.props.controlpanel.data}
               onSubmit={this.onSubmit}
               onCancel={this.onCancel}
