+++ src/config/ControlPanels.js
@@ -41,3 +41,61 @@ export const controlPanelsIcons = {
   undo: undoControlPanelSVG,
   aliases: linkSVG,
 };
+
+// Filters props.controlpanel.schema to only valid/relevant fields
+export const filterControlPanelsSchema = (controlpanel) => {
+  const panelType = controlpanel['@id'].split('/').pop();
+
+  const unwantedSettings = {
+    language: ['display_flags', 'always_show_selector'],
+    search: ['enable_livesearch'],
+    site: [
+      'display_publication_date_in_byline',
+      'icon_visibility',
+      'thumb_visibility',
+      'no_thumbs_portlet',
+      'no_thumbs_lists',
+      'no_thumbs_summary',
+      'no_thumbs_tables',
+      'thumb_scale_portlet',
+      'thumb_scale_listing',
+      'thumb_scale_table',
+      'thumb_scale_summary',
+      'toolbar_position',
+      'toolbar_logo',
+      'default_page',
+    ],
+    editing: ['available_editors', 'default_editor', 'ext_editor'],
+    imaging: [
+      'highpixeldensity_scales',
+      'quality_2x',
+      'quality_3x',
+      'picture_variants',
+      'image_captioning',
+    ],
+  };
+
+  // Creates modified version of properties object
+  const newPropertiesObj = Object.fromEntries(
+    Object.entries(controlpanel.schema.properties).filter(
+      ([key, val]) => !(unwantedSettings[panelType] || []).includes(key),
+    ),
+  );
+  // Filters props.controlpanel.schema.fieldsets.fields to only valid/relevant fields
+  const filterFields = (fields) => {
+    return fields.filter(
+      (field) => !(unwantedSettings[panelType] || []).includes(field),
+    );
+  };
+  // Creates modified version of fieldsets array
+  const newFieldsets = controlpanel.schema.fieldsets.map((fieldset) => {
+    return { ...fieldset, fields: filterFields(fieldset.fields) };
+  });
+
+  // Returns clone of props.controlpanel.schema, with updated properties/fieldsets
+  return {
+    ...controlpanel.schema,
+    properties: newPropertiesObj,
+    fieldsets: newFieldsets,
+  };
+};
